check for user input before deleting all records in database

+++ src/db/deleteAll.ts
@@ -1,6 +1,8 @@
 /* eslint-disable no-console */
 import dotenv from 'dotenv';
 import mongoose from 'mongoose';
+import { stdin as input, stdout as output } from 'node:process';
+import * as readline from 'node:readline/promises';
 
 import Comic from './models/Comic';
 import Illustration from './models/Illustration';
@@ -8,11 +10,23 @@ import Panel from './models/Panel';
 
 dotenv.config();
 
-const MONGODB_URI = process.env.MONGODB_URI || 'mongodb://localhost/ego_gala_comics';
-
-mongoose.connect(MONGODB_URI);
+const DEFAULT_MONGODB_URI = 'mongodb://localhost/ego_gala_comics';
+const MONGODB_URI = process.env.MONGODB_URI || DEFAULT_MONGODB_URI;
 
 const deleteAllRecords = async () => {
+  // protection against fat-fingering and deleting all records in
+  // a non-local database
+  if (MONGODB_URI !== DEFAULT_MONGODB_URI) {
+    const rl = readline.createInterface({ input, output });
+    const answer = await rl.question('Are you sure you want to delete all records? (y/n) ');
+    if (answer !== 'y') {
+      console.log('Aborting...');
+      process.exit(0);
+    }
+  }
+
+  mongoose.connect(MONGODB_URI);
+
   try {
     await Panel.deleteMany({});
     console.log('All Panel records deleted.');
@@ -30,4 +44,5 @@ const deleteAllRecords = async () => {
   }
 };
 
+console.log('Deleting all collections...');
 deleteAllRecords();
